+++ premium/backend/src/baserow_premium/api/admin/users/views.py
@@ -22,17 +22,17 @@ from baserow_premium.api.admin.views import AdminListingView
 from baserow_premium.license.handler import LicenseHandler
 from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import OpenApiParameter, extend_schema
+from rest_framework.generics import GenericAPIView
 from rest_framework.permissions import IsAdminUser
 from rest_framework.response import Response
-from rest_framework.status import HTTP_201_CREATED
+from rest_framework.status import HTTP_200_OK
 from rest_framework.views import APIView
-from rest_framework_jwt.authentication import JSONWebTokenAuthentication
-from rest_framework_jwt.serializers import ImpersonateAuthTokenSerializer
-from rest_framework_jwt.views import ImpersonateJSONWebTokenView
 
 from baserow.api.decorators import map_exceptions, validate_body
 from baserow.api.schemas import get_error_schema
 from baserow.api.user.schemas import authenticate_user_schema
+from baserow.api.user.serializers import get_all_user_data_serialized
+from baserow.core.user.utils import generate_session_tokens_for_user
 
 from .serializers import BaserowImpersonateAuthTokenSerializer
 
@@ -169,7 +169,17 @@ class UserAdminView(APIView):
         return Response(status=204)
 
 
-class UserAdminImpersonateView(ImpersonateJSONWebTokenView):
+class UserAdminImpersonateView(GenericAPIView):
+    """
+    Impersonate the user by retrieving its JWT.
+
+    Returns:
+        dict: {
+            "access_token": user's JWT token,
+            "refresh_refresh": user's JWT refresh token
+        }
+    """
+
     permission_classes = (IsAdminUser,)
     serializer_class = BaserowImpersonateAuthTokenSerializer
 
@@ -182,7 +192,7 @@ class UserAdminImpersonateView(ImpersonateJSONWebTokenView):
             "order to do this. It's not possible to impersonate a superuser or staff."
             "\n\nThis is a **premium** feature."
         ),
-        request=ImpersonateAuthTokenSerializer,
+        request=BaserowImpersonateAuthTokenSerializer,
         responses={
             200: authenticate_user_schema,
         },
@@ -199,16 +209,16 @@ class UserAdminImpersonateView(ImpersonateJSONWebTokenView):
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
-        token = serializer.validated_data.get("token")
-        user = serializer.validated_data.get("user")
-
-        response = JSONWebTokenAuthentication.jwt_create_response_payload(
-            token=token, user=user, request=request
-        )
+        user = serializer.validated_data["user"]
 
         logger.info(
             f"{request.user.username} ({request.user.id}) requested an "
             f"impersonate token for {user.username} ({user.id})."
         )
 
-        return Response(response, status=HTTP_201_CREATED)
+        serialized_data = {
+            **generate_session_tokens_for_user(user, include_refresh_token=True),
+            **get_all_user_data_serialized(user, request),
+        }
+
+        return Response(serialized_data, status=HTTP_200_OK)